requirements.txt
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
APScheduler==3.10.1
psycopg2==2.9.6
